requirements.txt
google-generativeai
python-dotenv
tiktoken
numpy
groq
requests
